Fixed DPI change issue in BandWindow.
Fixed an annoying issues where DPI changes weren't being perceived by BandWindow's HwndSource.

+++ ModernFlyouts.Core/Media/Source/SourceDesktopAppInfo.cs
@@ -35,7 +35,7 @@ namespace ModernFlyouts
                 {
                     if (GetWindowSizeState(hWnd) == WindowSizeState.Minimized)
                     {
-                        ShowWindowAsync(hWnd, (int)ShowWindowCommands.Restore);
+                        ShowWindowAsync(hWnd, ShowWindowCommands.Restore);
                     }
                     else
                     {
